requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
streamlit==1.38.0
google-generativeai==0.7.2
langchain==0.2.11
python-dotenv==1.0.1
pydantic==2.8.2
requests==2.32.3
reportlab==4.1.0
